% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_normalization_non_pair.R
\name{data_normalization_non_pair}
\alias{data_normalization_non_pair}
\title{data_normalization_non_pair}
\usage{
data_normalization_non_pair(sample_obj, selected_part)
}
\description{
data_normalization_non_pair
}
\examples{
checking.nominal()
}
