% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_normalization_pair.R
\name{data_normalization_pair}
\alias{data_normalization_pair}
\title{data_normalization_pair}
\usage{
data_normalization_pair(sample_obj, name)
}
\description{
data_normalization_pair
}
\examples{
checking.nominal()
}
