% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oversample_non_pair.R
\name{oversample_non_pair}
\alias{oversample_non_pair}
\title{oversample_non_pair}
\usage{
oversample_non_pair(bins, original_min_unit2 = 2,
  original_sample_unit2 = 200, by_factor = 1.1)
}
\description{
oversample_non_pair
}
