% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oversample_wrapper_non_pair.R
\name{oversample_wrapper_non_pair}
\alias{oversample_wrapper_non_pair}
\title{oversample_wrapper_non_pair}
\usage{
oversample_wrapper_non_pair(selected_part, random_number,
  original_min_red = 2, original_sample_red = 200, by_factor = 1.1)
}
\description{
oversample_wrapper_non_pair
}
