% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_100m_eyeballing.R
\name{prepare_100m_eyeballing}
\alias{prepare_100m_eyeballing}
\title{prepare_100m_eyeballing}
\usage{
prepare_100m_eyeballing(obj_100m)
}
\arguments{
\item{obj_100m}{SpatialPolygonsDataFrame}
}
\description{
prepare_100m_eyeballing
}
