% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_area.R
\name{prepare_area}
\alias{prepare_area}
\title{prepare_area}
\usage{
prepare_area(shape1, shape2, width_in_km = 100, split_width = NA)
}
\arguments{
\item{width_in_km}{in km}

\item{split_width}{in km}
}
\description{
prepare_area
}
