% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_raster_of_sampling_area.R
\name{prepare_raster_of_sampling_area}
\alias{prepare_raster_of_sampling_area}
\title{prepare_raster_of_sampling_area}
\usage{
prepare_raster_of_sampling_area(pop_raster, area_polygon)
}
\description{
prepare_raster_of_sampling_area
}
