% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_sampling_area.R
\name{prepare_sampling_area}
\alias{prepare_sampling_area}
\title{prepare_sampling_area}
\usage{
prepare_sampling_area(adm0, border_spatialline, lakes, width_in_km,
  split_width = NA)
}
\arguments{
\item{width_in_km}{in km}

\item{split_width}{in km}
}
\description{
prepare_sampling_area
}
