% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_from_raster.R
\name{sample_from_raster}
\alias{sample_from_raster}
\title{sample_from_raster}
\usage{
sample_from_raster(raster_file, poly_shape, sample_size = 20,
  seed = 1337, check_raster = NULL)
}
\description{
sample_from_raster
}
