% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_non_pair.R
\name{sample_non_pair}
\alias{sample_non_pair}
\title{sample_non_pair}
\usage{
sample_non_pair(sampling_bins, raster_1, raster_2, min_unit2 = 1,
  sample_unit2, random_number, verbose = 2)
}
\description{
sample_non_pair
}
