% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_pair.R
\name{sample_pair}
\alias{sample_pair}
\title{sample_pair}
\usage{
sample_pair(sampling_bins, raster_1, raster_2, raster_3, min_unit3 = 1,
  sample_unit3 = 100, sample_unit2 = 2, random_number, verbose = 2)
}
\description{
sample_pair
}
