% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_wrapper_non_pair_extract.R
\name{simulation_wrapper_non_pair_extract}
\alias{simulation_wrapper_non_pair_extract}
\title{simulation_wrapper_non_pair_extract}
\usage{
simulation_wrapper_non_pair_extract(selected_part, extract_list,
  random_number, original_min_red = 2, original_sample_red = 200,
  verbose = 2)
}
\arguments{
\item{extract_list}{List of validation rasters which are cut to the sampling bin outlines.}

\item{verbose}{If message should be printed. 0 no messages, 1 some, 2 all. 2 is default.}
}
\description{
simulation_wrapper_non_pair_extract
}
