% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oversample_wrapper_non_pair.R
\name{oversample_wrapper_non_pair}
\alias{oversample_wrapper_non_pair}
\title{oversample_wrapper_non_pair}
\usage{
oversample_wrapper_non_pair(name, bins, raster_1, raster_2, random_number,
  original_min_unit2 = 2, original_sample_unit2 = 200,
  by_factor = 1.1, verbose = 2)
}
\description{
oversample_wrapper_non_pair
}
