% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oversample_wrapper_pair.R
\name{oversample_wrapper_pair}
\alias{oversample_wrapper_pair}
\title{oversample_wrapper_pair}
\usage{
oversample_wrapper_pair(name, bins, raster_1, raster_2, raster_3,
  random_number, original_min_unit3 = 1, original_sample_unit3 = 200,
  by_factor = 1.1, verbose = 2)
}
\description{
oversample_wrapper_pair
}
