% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_non_pair.R
\name{sample_non_pair}
\alias{sample_non_pair}
\title{sample_non_pair}
\usage{
sample_non_pair(sampling_bins, red_raster_complete, yellow_raster_complete,
  min_red = 1, sample_red = 200, random_number)
}
\description{
sample_non_pair
}
