% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_pair.R
\name{sample_pair}
\alias{sample_pair}
\title{sample_pair}
\usage{
sample_pair(sampling_bins, blue_raster_complete, red_raster_complete,
  yellow_raster_complete, min_blue = 1, sample_blue = 100,
  sample_red = 2, random_number, verbose = 2)
}
\description{
sample_pair
}
